requirements.txt
fastapi
uvicorn
torch
scikit-learn
pillow
joblib
python-multipart
jinja2
